added download button

+++ module1/moduleOneHandleRoutines.c
@@ -165,9 +165,9 @@ LOCAL void _testModes(int routine_trigger)
     }
     else if ((uint32_t) READ_LINK_FAIL == test_case)
     {
-    	_diag_shm_ptr->rem_rcvr_cnt  = mdio_read_br(LINK_FAIL_COUNTER_REG) & REM_RCVR_CNT_MASK;
-    	_diag_shm_ptr->loc_rcvr_cnt  = mdio_read_br(LINK_FAIL_COUNTER_REG) & LOC_RCVR_CNT_MASK;
-    	_diag_shm_ptr->link_fail_cnt = mdio_read_br(GENERAL_STATUS_REG) & LINK_FAIL_CNT_MASK;
+    	_diag_shm_ptr->rem_rcvr_cnt  = (uint16_t) mdio_read_br(LINK_FAIL_COUNTER_REG) & REM_RCVR_CNT_MASK;
+    	_diag_shm_ptr->loc_rcvr_cnt  = (uint16_t) mdio_read_br(LINK_FAIL_COUNTER_REG) & LOC_RCVR_CNT_MASK;
+    	_diag_shm_ptr->link_fail_cnt = (uint16_t) mdio_read_br(GENERAL_STATUS_REG) & LINK_FAIL_CNT_MASK;
     }
     else
     {
